Declare CodeMirror as external variable
The current var CodeMirror statement is creating a new variable in this file that is never being set. Based on comments, it appears that the original intention is to access the global variable we get from pulling in the CodeMirror library. This updates it to use the syntax for declaring an external variable.